--without-shared-libs: Link Dovecot libraries with --whole-archive flag for binaries. This removes the need for the ugly unused_objects lists for binaries, which were needed to avoid plugins from failing because they were missing some functions.
Apparently there's no easy way to use --whole-archive properly with libtool,
so there's now a rather ugly cc-wrapper.sh that does it. Also this is done
only when GNU ld is deted. Most people are going to use --with-shared-libs
though, so I don't think any of this is going to be a real problem.